abora/Exercise5/reddit-api/requirements.txt
Django==2.1
djangorestframework==3.8.2
Pillow==5.2.0
pytz==2018.5
